* 'master' of github.com:miku/siskin:
  imslp: add note, refs #13055
  imslp: debug flag
  73: include marcbinary conversion
  docs: remove reqs.png, comments expire fast
  marburg: fix OAI endpoint

+++ siskin/sources/imslp.py
@@ -67,9 +67,18 @@ class IMSLPDownload(IMSLPTask):
 
 
 class IMSLPConvert(IMSLPTask):
-    """ Extract and transform. """
+    """
+    Extract and transform.
+
+    TODO, refs #13055.
+
+    File "/usr/lib/python2.7/site-packages/siskin/assets/15/15_marcbinary.py", line 165, in <module>
+        record = record["document"]
+    KeyError: 'document'
+    """
 
     date = ClosestDateParameter(default=datetime.date.today())
+    debug = luigi.BoolParameter(description='do not delete temporary folder')
 
     def requires(self):
         return IMSLPDownload(date=self.date)
@@ -79,7 +88,10 @@ class IMSLPConvert(IMSLPTask):
         shellout("tar -xzf {archive} -C {tempdir}", archive=self.input().path, tempdir=tempdir)
         output = shellout("python {script} {tempdir} {output}",
                           script=self.assets('15/15_marcbinary.py'), tempdir=tempdir)
-        shutil.rmtree(tempdir)
+        if not self.debug:
+            shutil.rmtree(tempdir)
+        else:
+            self.logger.debug("not deleting temporary folder at %s", tempdir)
         luigi.LocalTarget(output).move(self.output().path)
 
     def output(self):
